Add new button convention, refactor fetch into custom hook, fix loading message style, fix hover states, improve lighthouse score
