+++ src/App/App.css
@@ -19,7 +19,7 @@
 }
 
 .page-title {
-  font-size: 4vw;
+  font-size: 3.3vw;
   font-family: 'Playfair Display', serif;
   font-style: italic;
   margin-left: 1vw;
