+++ src/PainterInfo/PainterInfo.js
@@ -3,8 +3,8 @@ import { Link } from 'react-router-dom'
 import usePaintings from '../Hooks/usePaintings';
 import Gallery from '../Gallery/Gallery';
 import '../PaintingInfo/PaintingInfo.css'
-import backBtn from '../assets/back-btn.png'
-import tagBtn from '../assets/tagIcon.png'
+// import backBtn from '../assets/back-btn.png'
+// import tagBtn from '../assets/tagIcon.png'
 
 function PainterInfo(props) {
   let url;
@@ -23,10 +23,10 @@ function PainterInfo(props) {
   return (
     <section className="painter-page">
       <section className="painter-nav">
-        <Link to={"/"} style={{ textDecoration: 'none' }}>
-          <img src={backBtn} alt='back-btn' className='back-btn' />
+        <Link to={"/"} style={{ textDecoration: 'none', color: 'inherit' }}>
+          <h1 className="painting-page-title">ArtisTry</h1>
         </Link>
-        <h1 className="artist-name">{artistName}</h1>
+        <h1 className="artist-page-name">{artistName}</h1>
       </section>
       <section aria-label="gallery">
         <Gallery paintings={artistPaintings} setSelected={props.setSelected}/>
